Removed unused parameter from `after_create_path`
AFAIK the `resource` parameter for `resource_path` is only used if the value is a model instance with an id attribute. Here it's a model class instead, so it'll be ignored and the `resource_id` parameter is used instead. I think it'd be simpler and easier to understand if this unnecessary parameter was removed?