Pass resource to default
At the moment if a block is passed to default it is executed in isolation. It would be very useful to be able to access the model or the resource in default so that a default can be set that is based on parameters or other attributes (similar to how we can with visible).