enter submits rather than cancelling action
Pressing enter in a form will trigger the first button of type="submit"
in the form. This is the default type for buttons inside a form. Explicitly
set type="button" on the cancel button so that it won't be triggered
by enter. This causes the true submit button to be triggered instead.