Warn developers if they forget to set attachment_key on trix fields
To use attachments on a trix field, you need to set an attachment_key
in the field definition (which gives the name of the association
which attachments can be stored in).

Although there is code in the trix stimuls controller to warn the user
if they haven't set this, it doesn't work, because it expects the
`data-attachment-key` attribute to be set but empty and does not work
if the attribute isn't set at all (which is how the html actually works
if attachment_key isn't set).

Also, although the server throws an error if you try to attach something
without setting the attachment name (it throws an error because it cannot
find the right association to use), the error is unhelpful:

    TypeError ([...] is not a symbol nor a string)

(Where the array contains the name of every association on the model.)

I've updated this so that:

1. The javascript error is shown if the HTML attribute is empty OR undefined
2. It raises a more descriptive error on the server if it cannot find the correct association