Don't set the csrf token if it's not available
It's possible to disable csrf tokens by setting
`config.action_controller.allow_forgery_protection = false`
in your environment config file. This is the default
in the test environment.

Unfortunately, the trix field controller assumes that
the csrf token will always be available, which makes
it crash if it's not. Because of that it's not possible
to use avo trix fields in tests.

I've changed the code so that if the csrf token isn't
set then we don't use it.