spec: expect mapbox map to be rendered
This requires the ENV variable to be defined in the rspec command in
order for it to run correctly. If not, the API request will not be made,
and the canvas element will not be present.

However, having a 3rd party factor in our specs does not conform to best
practices. I had difficulty setting up VCR so I did not implement it
in this commit.

VCR could be set up, record a cassette, and then filter out the API_KEY
in the configuration settings, if desired.