Added zoom to location fields
Right now location fields always load at the default zoom level of 15 and there's no way to customize it.

I've added an extra attribute to the location field so you can override the zoom level in the resource. I've also made mapkick show the map controls so users can zoom in and out if they want to.