Correct how the tags helper gets the text for tags (#3398)
* Correct how the tags helper gets the text for tags

It was getting the `label` attribute of `.tagify__tag` elements before but so far as I can see `.tagify__tag` elements do not have a label. Because of that the tags helper was always returning an array of `nil` values (one for each tag).

Instead I've changed them to get the text of the tag, which is what  the `wait_for_tag_to_appear` and  `wait_for_tag_to_disappear` helpers do. This correctly returns the text of the tag.

It looks like this was not spotted before because the only test in the avo test suite that uses the tags helper just checks that an empty array of tags is returned.

* tags helper test

* fix test

* wait_for_loaded

* acts_as_ordered_taggable_on

---------

Co-authored-by: Paul Bob <paul.ionut.bob@gmail.com>
Co-authored-by: Paul Bob <69730720+Paul-Bob@users.noreply.github.com>