Fixed argument prefix warnings (#3594)
If I run an app using avo with warnings enabled I get a LOT of warnings like this:

    avo-3.16.1/app/components/avo/fields/files_field/edit_component.html.erb:1: warning: `**' interpreted as argument prefix

This is happening because a function is being called with `**` as the first argument but there are no brackets around the arguments. Adding brackets fixes the warning but it does not change the meaning, so I've done that for all field components.