Removed defer attribute from the avo.base.css stylesheet tag (#3635)
`defer` isn't a valid attribute for a `<link>` tag - it is for
`<script>` tags and having it here is invalid html. I think this is also
causing loading issues in google chrome for me when I redirect from avo
to my main app.

I've removed defer from the `<link>` tag. This should make the html
valid and shouldn't have any bad effects (as it shouldn't be used on a
`<link>` tag anyway).