chore: fix tests (#3898)
* chore: fix tests

* custom variant hover

I discovered that the Tailwind v4 hover change is affecting the test browser. As a temporary fix, I forced a custom variant to preserve the Tailwind v3 hover behavior, since it resolves the failing tests. However, we’ll need to find a more robust long-term solution this one was tricky to pinpoint.

* comment

* fix time spec

The field edit component for time field was rendering the hidden input as well, adding ! to hidden fixed it

* fix exposed_methods spec

* lint

* fix ket_Value

* fix skip_show_view_spec

* fix tags

tags element class was changed from hidden to !hidden but on the stimulus controller that change was not applied

* upload examples.txt

* try disable turbo on form

This is a temporary workaround

* re-add turbo on form

* fix reactive fields

* revert destroy_success_action